Rename TooltipV2 dev stories to examples for production deployment
